# openapi-chatgpt.yaml
openapi: 3.1.0
info:
  title: WeSign Digital Signature API
  description: Digital signature and document management with beautiful visual widgets
  version: 2.0.0
servers:
  - url: https://936da9734e85.ngrok-free.app
    description: WeSign MCP Server (ngrok tunnel)

paths:
  /execute:
    post:
      operationId: executeWeSignTool
      summary: Execute any WeSign tool
      description: Execute WeSign tools like listing documents, getting user info, managing templates, etc.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tool
              properties:
                tool:
                  type: string
                  description: "Tool name (e.g., wesign_get_user_info, wesign_list_documents, wesign_list_templates)"
                  enum:
                    - wesign_check_auth_status
                    - wesign_get_user_info
                    - wesign_list_documents
                    - wesign_list_templates
                    - wesign_get_document_info
                    - wesign_get_template
                    - wesign_search_documents
                    - wesign_upload_document
                    - wesign_create_template
                    - wesign_send_document_for_signing
                parameters:
                  type: object
                  description: "Tool-specific parameters (optional)"
                  additionalProperties: true
            examples:
              getUserInfo:
                summary: Get user account info
                value:
                  tool: wesign_get_user_info
                  parameters: {}
              listDocuments:
                summary: List documents
                value:
                  tool: wesign_list_documents
                  parameters:
                    offset: 0
                    limit: 10
              listTemplates:
                summary: List templates
                value:
                  tool: wesign_list_templates
                  parameters:
                    offset: 0
                    limit: 10
      responses:
        '200':
          description: Successful tool execution
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    description: Tool-specific response data
        '400':
          description: Invalid request
        '500':
          description: Server error

  /health:
    get:
      operationId: healthCheck
      summary: Check server health and authentication status
      description: Returns server status and authentication info
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  authenticated:
                    type: boolean
                  protocol:
                    type: string
                  timestamp:
                    type: string
                    format: date-time

components:
  schemas:
    ToolRequest:
      type: object
      required:
        - tool
      properties:
        tool:
          type: string
          description: Name of the WeSign tool to execute
        parameters:
          type: object
          description: Tool-specific parameters

    ToolResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string